Fix a bug in the bp percentile
Prior to this change the calculated height percentiles were in the range
(0, 1) in the cpp code.  The look up table expected values [0, 100] and
thus the look up always picked the 5th percentile for height leading to
elevated blood pressure precentiles.

This change adds a needed "100 *" to the return from the cppPGSF1 call
within the cppBP call.

A test has been added to the package checking the specific data example
that identified that something was wrong.